Add files via upload
Logical.cpp is a logical clock based on Lamport's algorithm. The program read in a set of simulation
parameters and then print the logical clocks associated with each event.
Vector.cpp is a vector clock which has been implemented using Message Passing Interface(MPI).
The code implement is set of primitives to maintain the state of vector clocks. This programs read in a set of simulation
parameters and then print the vector clocks associated with each event. It is recommended that a
single manager process and a set of worker processes be implemented.